Rename crates in order to make LibNyx publishable
