+++ acat/Cargo.toml
@@ -7,6 +7,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-fuzz_runner={path="../fuzz_runner"}
+libnyx_fuzz_runner={path="../libnyx_fuzz_runner"}
 colored = "2.0.0"
 clap="2.33.0"
