+++ libnyx/src/ffi.rs
@@ -19,7 +19,7 @@ use libc::c_char;
 use std::ffi::CStr;
 use std::ffi::c_void;
 
-use fuzz_runner::nyx::aux_buffer::{NYX_CRASH, NYX_HPRINTF, NYX_ABORT};
+use libnyx_fuzz_runner::nyx::aux_buffer::{NYX_CRASH, NYX_HPRINTF, NYX_ABORT};
 use super::*;
 
 /* Helper function to load a C string pointer and return a Rust string. */
