+++ libnyx_fuzz_runner/src/nyx/aux_buffer.rs
