+++ libnyx_fuzz_runner/src/nyx/ijon_data.rs
