+++ libnyx_fuzz_runner/src/nyx/mod.rs
@@ -9,7 +9,7 @@ pub use qemu_process::QemuProcess;
 use std::fs;
 use std::path::PathBuf;
 
-extern crate config;
+extern crate libnyx_config;
 
 fn into_absolute_path(sharedir: &str) -> String{
 
@@ -23,7 +23,7 @@ fn into_absolute_path(sharedir: &str) -> String{
     }
 }
 
-pub fn qemu_process_new(sharedir: String, cfg: &config::Config) -> Result<QemuProcess, String> {
+pub fn qemu_process_new(sharedir: String, cfg: &libnyx_config::Config) -> Result<QemuProcess, String> {
 
 
     let qemu_params = params::QemuParams::new(into_absolute_path(&sharedir), cfg);
